T5-269 Remove named exports
