+++ src/components/Button/index.js
@@ -1,3 +1,1 @@
 export { default } from './Button';
-
-export { default as FakeButton } from './FakeButton';
