+++ src/components/NewsletterForm/NewsletterOverlayForm/NewsletterOverlayForm.js
@@ -1,13 +1,15 @@
 import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import { css } from '@emotion/react';
 import loadable from '@loadable/component';
 
-import { encode } from '../../util';
+import { encode } from '../../../util';
+import { colors } from '../../../styles';
 
-const Input = loadable(() => import('../Input'));
-const Button = loadable(() => import('../Button'));
+const Input = loadable(() => import('../../Input'));
+const Button = loadable(() => import('../../Button'));
 
-export default () => {
+const NewsletterOverlayForm = ({ setIsActive, isActive }) => {
   const [email, updateEmail] = useState('');
   const [submitted, hasSubmitted] = useState(false);
 
@@ -24,6 +26,9 @@ export default () => {
     }).then(() => {
       updateEmail('');
       hasSubmitted(true);
+      setTimeout(() => {
+        setIsActive(!isActive);
+      }, 1000);
     });
   };
 
@@ -54,7 +59,7 @@ export default () => {
         </legend>
         {/* eslint-disable-next-line */}
         <label
-          htmlFor='nws-email'
+          htmlFor='nws-emailoverlay'
           css={css`
             display: none;
           `}
@@ -63,18 +68,37 @@ export default () => {
         </label>
         {!submitted && (
           <Input
+            css={css`
+              background-color: ${colors.white};
+              letter-spacing: 0;
+            `}
             type='email'
             name='email'
-            id='nws-email'
+            id='nws-emailoverlay'
             placeholder='Email'
             value={email}
             onChange={evt => updateEmail(evt.target.value)}
           />
         )}
       </fieldset>
-      <Button onClick={onSubmit} disabled={submitted}>
+      <Button
+        onClick={onSubmit}
+        disabled={submitted}
+        css={css`
+          max-width: 200px;
+          width: 100%;
+          margin: auto;
+        `}
+      >
         {submitted ? 'Thank You' : 'Sign Me Up'}
       </Button>
     </form>
   );
 };
+
+NewsletterOverlayForm.propTypes = {
+  setIsActive: PropTypes.func.isRequired,
+  isActive: PropTypes.bool.isRequired,
+};
+
+export default NewsletterOverlayForm;
