+++ src/components/NewsletterForm/NewsletterOverlayForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './NewsletterOverlayForm';
