+++ src/components/NewsletterForm/OverlayForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './OverlayForm';
