--- src/components/NewsletterForm/index.js
@@ -1,7 +1,0 @@
-export { default } from './NewsletterForm';
-export { default as NewsletterFullWidthSection } from './NewsletterFullWidthSection';
-export { default as NewsletterOverlay } from './NewsletterOverlay';
-export { default as NewsletterOverlayForm } from './NewsletterOverlayForm';
-export { default as NewsletterSimpleOverlay } from './NewsletterSimpleOverlay';
-export { default as NewsletterSimpleOverlayForm } from './NewsletterSimpleOverlayForm';
-export { default as OverlayForm } from './OverlayForm';
