+++ src/components/Prefooter/GetInTouch/GetInTouch.js
@@ -3,10 +3,10 @@ import { navigate } from 'gatsby';
 import { css } from '@emotion/react';
 import loadable from '@loadable/component';
 
-import { colors, mediaQueries } from '../../styles';
+import { colors, mediaQueries } from '../../../styles';
 
-const Button = loadable(() => import('../Button'));
-const TextWrapper = loadable(() => import('./TextWrapper'));
+const Button = loadable(() => import('../../Button'));
+const TextWrapper = loadable(() => import('../TextWrapper'));
 
 const wrapperStyles = css`
   ${mediaQueries.phoneLarge} {
@@ -14,11 +14,11 @@ const wrapperStyles = css`
   }
 `;
 
-const ContactUs = () => (
+const GetInTouch = () => (
   <TextWrapper backgroundColor={colors.yellow} css={wrapperStyles}>
-    <h3>Get to know us.</h3>
+    <h3>Get in touch today.</h3>
     <Button onClick={() => navigate(`/contact/`)}>Contact Us</Button>
   </TextWrapper>
 );
 
-export default ContactUs;
+export default GetInTouch;
