+++ src/components/Prefooter/SeeInsights/index.js
@@ -1,0 +1,1 @@
+export { default } from './SeeInsights';
