+++ src/components/Prefooter/TextWrapper/index.js
@@ -1,0 +1,1 @@
+export { default } from './TextWrapper';
