+++ src/components/Prefooter/WhatWeDo/WhatWeDo.js
@@ -4,10 +4,10 @@ import { navigate } from 'gatsby';
 import { css } from '@emotion/react';
 import loadable from '@loadable/component';
 
-import { colors, mediaQueries } from '../../styles';
+import { colors, mediaQueries } from '../../../styles';
 
-const Button = loadable(() => import('../Button'));
-const TextWrapper = loadable(() => import('./TextWrapper'));
+const Button = loadable(() => import('../../Button'));
+const TextWrapper = loadable(() => import('../TextWrapper'));
 
 const wrapperStyles = css`
   ${mediaQueries.phoneLarge} {
