--- src/components/Prefooter/index.js
@@ -1,6 +1,0 @@
-export { default as TextWrapper } from './TextWrapper';
-export { default as BeUs } from './BeUs';
-export { default as ContactUs } from './ContactUs';
-export { default as SeeInsights } from './SeeInsights';
-export { default as GetInTouch } from './GetInTouch';
-export { default as WhatWeDo } from './WhatWeDo';
